factor out swipe action into convenience method
