+++ EcoSoapBank/Helpers/UIKit/UISwipeActionsConfiguration+Constants.swift
@@ -6,4 +6,21 @@
 //  Copyright © 2020 Spencer Curtis. All rights reserved.
 //
 
-import Foundation
+import UIKit
+
+
+extension UISwipeActionsConfiguration {
+    static func remove(_ deleteHandler: @escaping UIContextualAction.Handler
+    ) -> UISwipeActionsConfiguration {
+        let delete = UIContextualAction(
+            style: .destructive,
+            title: "Remove",
+            handler: deleteHandler)
+        delete.backgroundColor = .systemRed
+
+        let config = UISwipeActionsConfiguration(actions: [delete])
+        config.performsFirstActionWithFullSwipe = true
+
+        return config
+    }
+}
